+++ src/main/java/org/torproject/onionoo/writer/ClientsDocumentWriter.java
@@ -78,22 +78,16 @@ public class ClientsDocumentWriter implements DocumentWriter {
   }
 
   private String[] graphNames = new String[] {
-      "1_week",
-      "1_month",
       "3_months",
       "1_year",
       "5_years" };
 
   private Period[] graphIntervals = new Period[] {
-      Period.ofWeeks(1),
-      Period.ofMonths(1),
       Period.ofMonths(3),
       Period.ofYears(1),
       Period.ofYears(5) };
 
   private long[] dataPointIntervals = new long[] {
-      DateTimeHelper.ONE_DAY,
-      DateTimeHelper.ONE_DAY,
       DateTimeHelper.ONE_DAY,
       DateTimeHelper.TWO_DAYS,
       DateTimeHelper.TEN_DAYS };