+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
  * Medium changes
    - Extend "version" parameter to support lists and ranges.
+   - Remove redundant "1_week" and "1_month" graphs from clients
+     documents.
 
 
 # Changes in version 6.2-1.17.1 - 2018-08-17